Bugfix for fetching new refs misses some commits (#6)
I think there's a bug with the old code where if you don't run fetch in the repo before running deploy, then config.stagingRef and config.productionRef might be wrong, because the upstream may have been updated. This PR changes the approach to use github API for comparing commits, instead of replying on the local repo.